Refactor: move dataclass predicate to `utils/dataclass.py`
