+++ tests/utils/test_dataclass.py
@@ -5,7 +5,12 @@ from dataclasses import dataclass
 import pytest
 
 from litestar.types import Empty, EmptyType
-from litestar.utils.dataclass import extract_dataclass_fields, extract_dataclass_items, simple_asdict
+from litestar.utils.dataclass import (
+    extract_dataclass_fields,
+    extract_dataclass_items,
+    is_dataclass_class,
+    simple_asdict,
+)
 
 
 def test_extract_dataclass_fields_exclude_none() -> None:
@@ -141,3 +146,16 @@ def test_simple_asdict_does_not_recurse_into_collections() -> None:
     foo = Foo()
 
     assert simple_asdict(Bar(foo=[foo])) == {"foo": [foo]}
+
+
+def test_is_dataclass_class() -> None:
+    """is_dataclass_class() should return True for types and False for instances."""
+
+    @dataclass
+    class Foo:
+        """A Foo model."""
+
+        bar: str = "bar"
+
+    assert is_dataclass_class(Foo)
+    assert not is_dataclass_class(Foo())
